Refactored initialization code
