+++ public/index.html
@@ -191,7 +191,7 @@
                     <div class="justify-flexbox">
                         <img class="appearanceIcon" id="Aqua" src="images/modal_boxes/puyo_aqua.png">
                         <img class="appearanceIcon" id="Chalk" src="images/modal_boxes/puyo_chalk.png">
-                        <img class="appearanceIcon selected" id="TsuClassic" src="images/modal_boxes/puyo_tsu.png">
+                        <img class="appearanceIcon" id="TsuClassic" src="images/modal_boxes/puyo_tsu.png">
                         <img class="appearanceIcon" id="Custom" src="images/modal_boxes/puyo_custom.png">
                         <img class="appearanceIcon" id="FlatColour" src="images/modal_boxes/puyo_flat.png">
                         <img class="appearanceIcon" id="Test" src="images/modal_boxes/puyo_teal.png">
